faster-whisper-webui/requirements-whisper.txt
git+https://github.com/huggingface/transformers
git+https://github.com/openai/whisper.git
transformers
ffmpeg-python==0.2.0
gradio==3.23.0
yt-dlp
torchaudio
altair
json5